distance and force distance functions updated to always reflect current top distance + always make the step as long as needed + debug info for memory leaks added

+++ projects/include/FIFO.h
@@ -9,7 +9,7 @@ struct Coord
   Coord *nextCoord;
   bool intersected = false;
   int depth = 0;
-  int distance_at_point = 0;
+  float distance_at_point = 0.0f;
 };
 
 void InsertCoord(Coord **stack, int x, int y);
